+++ Semana_3/gobarber/src/pages/SignIn/index.tsx
@@ -24,7 +24,7 @@ const SignIn: React.FC = () => (
         <a href="forgot">Esqueci minha senha</a>
       </form>
 
-      <a href="forgot">
+      <a href="/SignUp">
         <FiLogIn />
         Criar Conta
       </a>